requirements.txt
captcha
Pillow